requirements.txt
botcity-framework-web==0.9.0
botcity-maestro-sdk>=0.3.3,<1.0

botcity-utils>=0.1.2

pytesseract
keyboard==0.13.5
MouseInfo==0.1.3
numpy==1.26.4
opencv-python==4.9.0.80
# pillow==10.3.0
PyAutoGUI==0.9.54
setuptools
# PyGetWindow==0.0.9
# PyMsgBox==1.0.9
# pynput==1.7.7
# pyperclip==1.8.2
# PyRect==0.2.0
# PyScreeze==0.1.30
# pytweening==1.2.0
# six==1.16.0

# Uncomment to use Pandas
pandas
